Handle image and alt text data variations in source
This commit supports various data shapes for the product's image and alt text. It can now handle sources represented as strings or as objects containing a 'url' attribute. Changes were made in cloudinary helper, product grid and product detail modules. Also, empty alt texts are now taken from the title attribute if not given.

In addition, local storage availability was checked in base-http service. Add a method to update shipping in the cart service and subscribe to the cart for detecting the shipping method in the checkout process. All these changes improve the overall robustness and flexibility of the code.